+++ packages/porygon-typeorm/src/crud/modules/iocuak/TypeOrmReadContainerModule.spec.ts
@@ -1,6 +1,11 @@
 jest.mock('@cuaklabs/iocuak');
 
-import { Container, inject, injectable, BindingScope } from '@cuaklabs/iocuak';
+import {
+  inject,
+  injectable,
+  BindingScope,
+  ContainerModuleBindingService,
+} from '@cuaklabs/iocuak';
 import { CrudModuleType, ModuleTypeToSymbolMap } from '@cuaklabs/porygon';
 
 import { FindTypeOrmAdapter } from '../../adapter/typeorm/FindTypeOrmAdapter';
@@ -63,12 +68,14 @@ describe(TypeOrmReadContainerModule.name, () => {
   });
 
   describe('.load()', () => {
-    let containerApiMock: jest.Mocked<Container>;
+    let containerModuleBindingServiceMock: jest.Mocked<ContainerModuleBindingService>;
 
     beforeAll(() => {
-      containerApiMock = {
+      containerModuleBindingServiceMock = {
         bind: jest.fn(),
-      } as Partial<jest.Mocked<Container>> as jest.Mocked<Container>;
+      } as Partial<
+        jest.Mocked<ContainerModuleBindingService>
+      > as jest.Mocked<ContainerModuleBindingService>;
     });
 
     describe('when called', () => {
@@ -99,7 +106,7 @@ describe(TypeOrmReadContainerModule.name, () => {
           injectableDecoratorMock,
         );
 
-        typeOrmReadContainerModule.load(containerApiMock);
+        typeOrmReadContainerModule.load(containerModuleBindingServiceMock);
       });
 
       afterAll(() => {
@@ -172,8 +179,8 @@ describe(TypeOrmReadContainerModule.name, () => {
       });
 
       it('should call containerApi.bind()', () => {
-        expect(containerApiMock.bind).toHaveBeenCalledTimes(1);
-        expect(containerApiMock.bind).toHaveBeenCalledWith(
+        expect(containerModuleBindingServiceMock.bind).toHaveBeenCalledTimes(1);
+        expect(containerModuleBindingServiceMock.bind).toHaveBeenCalledWith(
           expectClassExtending(FindTypeOrmAdapter),
         );
       });