+++ packages/porygon/src/crud/modules/iocuak/DomainCrudContainerModule.spec.ts
@@ -3,7 +3,10 @@ jest.mock('./DomainDeleteContainerModule');
 jest.mock('./DomainReadContainerModule');
 jest.mock('./DomainUpdateContainerModule');
 
-import { Container } from '@cuaklabs/iocuak';
+import {
+  ContainerModule,
+  ContainerModuleBindingService,
+} from '@cuaklabs/iocuak';
 
 import { CrudModuleType } from '../../models/domain/CrudModuleType';
 import { ModuleTypeToSymbolMap } from '../../models/domain/ModuleTypeToSymbolMap';
@@ -13,49 +16,25 @@ import { DomainDeleteContainerModule } from './DomainDeleteContainerModule';
 import { DomainReadContainerModule } from './DomainReadContainerModule';
 import { DomainUpdateContainerModule } from './DomainUpdateContainerModule';
 
-interface ModelTest {
-  foo: string;
-}
-
-interface QueryTest {
-  bar: string;
-}
-
 describe(DomainCrudContainerModule.name, () => {
-  let domainCreationContainerModuleMock: jest.Mocked<
-    DomainCreateContainerModule<ModelTest, QueryTest>
-  >;
-  let domainDeleteContainerModuleMock: jest.Mocked<
-    DomainDeleteContainerModule<QueryTest>
-  >;
-  let domainReadContainerModuleMock: jest.Mocked<
-    DomainReadContainerModule<ModelTest, QueryTest>
-  >;
-  let domainUpdateContainerModuleMock: jest.Mocked<
-    DomainUpdateContainerModule<QueryTest>
-  >;
+  let domainCreationContainerModuleMock: jest.Mocked<ContainerModule>;
+  let domainDeleteContainerModuleMock: jest.Mocked<ContainerModule>;
+  let domainReadContainerModuleMock: jest.Mocked<ContainerModule>;
+  let domainUpdateContainerModuleMock: jest.Mocked<ContainerModule>;
 
   beforeAll(() => {
     domainCreationContainerModuleMock = {
       load: jest.fn(),
-    } as Partial<
-      jest.Mocked<DomainCreateContainerModule<ModelTest, QueryTest>>
-    > as jest.Mocked<DomainCreateContainerModule<ModelTest, QueryTest>>;
+    } as Partial<jest.Mocked<ContainerModule>> as jest.Mocked<ContainerModule>;
     domainDeleteContainerModuleMock = {
       load: jest.fn(),
-    } as Partial<
-      jest.Mocked<DomainDeleteContainerModule<QueryTest>>
-    > as jest.Mocked<DomainDeleteContainerModule<QueryTest>>;
+    } as Partial<jest.Mocked<ContainerModule>> as jest.Mocked<ContainerModule>;
     domainReadContainerModuleMock = {
       load: jest.fn(),
-    } as Partial<
-      jest.Mocked<DomainReadContainerModule<ModelTest, QueryTest>>
-    > as jest.Mocked<DomainReadContainerModule<ModelTest, QueryTest>>;
+    } as Partial<jest.Mocked<ContainerModule>> as jest.Mocked<ContainerModule>;
     domainUpdateContainerModuleMock = {
       load: jest.fn(),
-    } as Partial<
-      jest.Mocked<DomainUpdateContainerModule<QueryTest>>
-    > as jest.Mocked<DomainUpdateContainerModule<QueryTest>>;
+    } as Partial<jest.Mocked<ContainerModule>> as jest.Mocked<ContainerModule>;
 
     (DomainCreateContainerModule as jest.Mock).mockReturnValue(
       domainCreationContainerModuleMock,
@@ -73,10 +52,7 @@ describe(DomainCrudContainerModule.name, () => {
 
   describe('when instantiated', () => {
     let crudModuleTypeToSymbolMap: ModuleTypeToSymbolMap<CrudModuleType>;
-    let domainCrudContainerModule: DomainCrudContainerModule<
-      ModelTest,
-      QueryTest
-    >;
+    let domainCrudContainerModule: DomainCrudContainerModule;
 
     beforeAll(() => {
       crudModuleTypeToSymbolMap = Object.freeze({
@@ -126,14 +102,16 @@ describe(DomainCrudContainerModule.name, () => {
 
     describe('.load()', () => {
       describe('when called', () => {
-        let containerApiMock: jest.Mocked<Container>;
+        let containerModuleBindingServiceMock: jest.Mocked<ContainerModuleBindingService>;
 
         beforeAll(() => {
-          containerApiMock = {
+          containerModuleBindingServiceMock = {
             bind: jest.fn(),
-          } as Partial<jest.Mocked<Container>> as jest.Mocked<Container>;
+          } as Partial<
+            jest.Mocked<ContainerModuleBindingService>
+          > as jest.Mocked<ContainerModuleBindingService>;
 
-          domainCrudContainerModule.load(containerApiMock);
+          domainCrudContainerModule.load(containerModuleBindingServiceMock);
         });
 
         afterAll(() => {
@@ -145,28 +123,28 @@ describe(DomainCrudContainerModule.name, () => {
             1,
           );
           expect(domainCreationContainerModuleMock.load).toHaveBeenCalledWith(
-            containerApiMock,
+            containerModuleBindingServiceMock,
           );
         });
 
         it('should call DomainDeleteContainerModule.load()', () => {
           expect(domainDeleteContainerModuleMock.load).toHaveBeenCalledTimes(1);
           expect(domainDeleteContainerModuleMock.load).toHaveBeenCalledWith(
-            containerApiMock,
+            containerModuleBindingServiceMock,
           );
         });
 
         it('should call DomainReadContainerModule.load()', () => {
           expect(domainReadContainerModuleMock.load).toHaveBeenCalledTimes(1);
           expect(domainReadContainerModuleMock.load).toHaveBeenCalledWith(
-            containerApiMock,
+            containerModuleBindingServiceMock,
           );
         });
 
         it('should call DomainUpdateContainerModule.load()', () => {
           expect(domainUpdateContainerModuleMock.load).toHaveBeenCalledTimes(1);
           expect(domainUpdateContainerModuleMock.load).toHaveBeenCalledWith(
-            containerApiMock,
+            containerModuleBindingServiceMock,
           );
         });
       });