+++ packages/porygon/src/crud/modules/iocuak/DomainUpdateContainerModule.spec.ts
@@ -1,6 +1,11 @@
 jest.mock('@cuaklabs/iocuak');
 
-import { Container, inject, injectable, BindingScope } from '@cuaklabs/iocuak';
+import {
+  inject,
+  injectable,
+  BindingScope,
+  ContainerModuleBindingService,
+} from '@cuaklabs/iocuak';
 
 import { CrudModuleType } from '../../models/domain/CrudModuleType';
 import { ModuleTypeToSymbolMap } from '../../models/domain/ModuleTypeToSymbolMap';
@@ -44,12 +49,14 @@ describe(DomainUpdateContainerModule.name, () => {
   });
 
   describe('.load()', () => {
-    let containerApiMock: jest.Mocked<Container>;
+    let containerModuleBindingServiceMock: jest.Mocked<ContainerModuleBindingService>;
 
     beforeAll(() => {
-      containerApiMock = {
+      containerModuleBindingServiceMock = {
         bind: jest.fn(),
-      } as Partial<jest.Mocked<Container>> as jest.Mocked<Container>;
+      } as Partial<
+        jest.Mocked<ContainerModuleBindingService>
+      > as jest.Mocked<ContainerModuleBindingService>;
     });
 
     describe('when called', () => {
@@ -79,7 +86,7 @@ describe(DomainUpdateContainerModule.name, () => {
           injectableDecoratorMock,
         );
 
-        domainUpdateContainerModule.load(containerApiMock);
+        domainUpdateContainerModule.load(containerModuleBindingServiceMock);
       });
 
       afterAll(() => {
@@ -116,8 +123,8 @@ describe(DomainUpdateContainerModule.name, () => {
       });
 
       it('should call containerApi.bind()', () => {
-        expect(containerApiMock.bind).toHaveBeenCalledTimes(1);
-        expect(containerApiMock.bind).toHaveBeenCalledWith(
+        expect(containerModuleBindingServiceMock.bind).toHaveBeenCalledTimes(1);
+        expect(containerModuleBindingServiceMock.bind).toHaveBeenCalledWith(
           expectClassExtending(UpdateEntityInteractor),
         );
       });